+++ src/__mocks__/mocks.js
@@ -5,7 +5,7 @@ import utils from '../utils';
  *
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  */
-export function mockGetEntryNames (chunksWebpackPlugin, entryNames) {
+export function mockGetEntryNames(chunksWebpackPlugin, entryNames) {
 	chunksWebpackPlugin.getEntryNames = jest.fn().mockImplementation(() => {
 		return entryNames;
 	});
@@ -17,7 +17,7 @@ export function mockGetEntryNames (chunksWebpackPlugin, entryNames) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Array} files List of entrypoint names
  */
-export function mockGetFiles (chunksWebpackPlugin, files) {
+export function mockGetFiles(chunksWebpackPlugin, files) {
 	chunksWebpackPlugin.getFiles = jest.fn().mockImplementation(() => {
 		return files;
 	});
@@ -29,7 +29,7 @@ export function mockGetFiles (chunksWebpackPlugin, files) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Boolean} status Status of the return of the function
  */
-export function mockHasCustomFormatTags (chunksWebpackPlugin, status) {
+export function mockHasCustomFormatTags(chunksWebpackPlugin, status) {
 	chunksWebpackPlugin.hasCustomFormatTags = jest.fn().mockImplementation(() => {
 		return status;
 	});
@@ -41,7 +41,7 @@ export function mockHasCustomFormatTags (chunksWebpackPlugin, status) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Object} htmlTags HTML tags with all assets for an entrypoint and sorted by type
  */
-export function mockGetHtmlTags (chunksWebpackPlugin, htmlTags) {
+export function mockGetHtmlTags(chunksWebpackPlugin, htmlTags) {
 	chunksWebpackPlugin.getHtmlTags = jest.fn().mockImplementation(() => {
 		return htmlTags;
 	});
@@ -53,7 +53,7 @@ export function mockGetHtmlTags (chunksWebpackPlugin, htmlTags) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Object} htmlTags HTML tags with all assets for an entrypoint and sorted by type
  */
-export function mockCustomFormatTags (chunksWebpackPlugin, htmlTags) {
+export function mockCustomFormatTags(chunksWebpackPlugin, htmlTags) {
 	chunksWebpackPlugin.options.customFormatTags = jest.fn().mockImplementation(() => {
 		return htmlTags;
 	});
@@ -65,7 +65,7 @@ export function mockCustomFormatTags (chunksWebpackPlugin, htmlTags) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Boolean} status Status of the return of the function
  */
-export function mockIsValidCustomFormatTagsDatas (chunksWebpackPlugin, status) {
+export function mockIsValidCustomFormatTagsDatas(chunksWebpackPlugin, status) {
 	chunksWebpackPlugin.isValidCustomFormatTagsDatas = jest.fn().mockImplementation(() => {
 		return status;
 	});
@@ -77,7 +77,7 @@ export function mockIsValidCustomFormatTagsDatas (chunksWebpackPlugin, status) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Object} chunks All chunks sorted by extension type
  */
-export function mockSortsChunksByType (chunksWebpackPlugin, chunks) {
+export function mockSortsChunksByType(chunksWebpackPlugin, chunks) {
 	chunksWebpackPlugin.sortsChunksByType = jest.fn().mockImplementation(() => {
 		return chunks;
 	});
@@ -89,7 +89,7 @@ export function mockSortsChunksByType (chunksWebpackPlugin, chunks) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Boolean} status Status of the return of the function
  */
-export function mockIsPublicPathNeedsEndingSlash (chunksWebpackPlugin, status) {
+export function mockIsPublicPathNeedsEndingSlash(chunksWebpackPlugin, status) {
 	chunksWebpackPlugin.isPublicPathNeedsEndingSlash = jest.fn().mockImplementation(() => {
 		return status;
 	});
@@ -101,7 +101,7 @@ export function mockIsPublicPathNeedsEndingSlash (chunksWebpackPlugin, status) {
  * @param {Class} chunksWebpackPlugin Instance of chunksWebpackPlugin
  * @param {Boolean} status Status of the return of the function
  */
-export function mockIsValidOutputPath (chunksWebpackPlugin, status) {
+export function mockIsValidOutputPath(chunksWebpackPlugin, status) {
 	chunksWebpackPlugin.isValidOutputPath = jest.fn().mockImplementation(() => {
 		return status;
 	});
@@ -112,7 +112,7 @@ export function mockIsValidOutputPath (chunksWebpackPlugin, status) {
  *
  * @param {Boolean} status Status of the return of the function
  */
-export function mockIsAbsolutePath (status) {
+export function mockIsAbsolutePath(status) {
 	utils.isAbsolutePath = jest.fn().mockImplementation(() => {
 		return status;
 	});