+++ src/__tests__/index.test.js
@@ -31,12 +31,12 @@ const options = {
 	customFormatTags: (chunksSorted, files) => {
 		// Generate all HTML style tags with CDN prefix
 		const styles = chunksSorted.styles
-			.map(chunkCss => `<link rel="stylesheet" href="https://cdn.domain.com${chunkCss}" />`)
+			.map((chunkCss) => `<link rel="stylesheet" href="https://cdn.domain.com${chunkCss}" />`)
 			.join('');
 
 		// Generate all HTML style tags with CDN prefix and defer attribute
 		const scripts = chunksSorted.scripts
-			.map(chunkJs => `<script defer src="https://cdn.domain.com${chunkJs}"></script>`)
+			.map((chunkJs) => `<script defer src="https://cdn.domain.com${chunkJs}"></script>`)
 			.join('');
 
 		return { styles, scripts };