+++ src/utils.ts
@@ -1,5 +1,11 @@
 import path = require('path');
-import fse = require('fs-extra');
+
+// Describe the shape of the Fs object
+interface Fs {
+	existsSync: Function;
+	mkdirSync: Function;
+	writeFileSync: Function;
+}
 
 /**
  * Throw an error
@@ -33,11 +39,17 @@ const getFileExtension = (file: string): string => {
 /**
  * Write content into a file with a specific output path
  *
+ * @param {Object} fs The input file system from the webpack compiler
  * @param {String} outputPath The output path of the file
  * @param {String} output The output of the file
  */
-const writeFile = ({ outputPath, output }: { outputPath: string; output: string }) => {
-	fse.outputFileSync(outputPath, output);
+const writeFile = ({ fs, outputPath, output }: { fs: Fs; outputPath: string; output: string }) => {
+	// Check if the path contains missing directories, else we create them recursively
+	const dir = path.dirname(outputPath);
+	if (!fs.existsSync(dir)) {
+		fs.mkdirSync(dir, { recursive: true });
+	}
+	fs.writeFileSync(outputPath, output);
 };
 
 /**